refactor(extensions): network extension
