+++ src/extensions/network/capsule-builder.ts
@@ -5,9 +5,9 @@ import hash from 'object-hash';
 import v4 from 'uuid';
 import filenamify from 'filenamify';
 import { BitId } from '../../bit-id';
-import orchestrator, { CapsuleOrchestrator } from '../capsule/orchestrator/orchestrator';
+import orchestrator, { CapsuleOrchestrator } from './orchestrator/orchestrator';
 import { PackageManager } from '../package-manager';
-import { CapsuleOptions, CreateOptions } from '../capsule/orchestrator/types';
+import { CapsuleOptions, CreateOptions } from './orchestrator/types';
 import Consumer from '../../consumer/consumer';
 import { ComponentCapsule } from '../capsule-ext';
 import { getComponentLinks } from '../../links/link-generator';
