+++ src/extensions/network/capsule-list.ts
@@ -1,0 +1,4 @@
+import BitIdAndValueArray from '../../bit-id/bit-id-and-value-array';
+import { ComponentCapsule } from '../capsule-ext';
+
+export default class CapsuleList extends BitIdAndValueArray<ComponentCapsule> {}
