+++ src/extensions/network/index.ts
@@ -1,0 +1,2 @@
+export { default as NetworkExt } from './network.manifest';
+export { default as Network } from './network';
