+++ src/extensions/network/network.manifest.ts
@@ -1,0 +1,9 @@
+import Network from './network';
+import { Capsule } from '../capsule';
+
+export default {
+  name: 'Network',
+  dependencies: [Capsule],
+  config: {},
+  provider: Network.provide
+};
