+++ src/extensions/network/orchestrator/resource-pool/resource-factory.ts
