+++ src/extensions/network/types.ts
@@ -1,0 +1,6 @@
+import { ComponentCapsule } from '../capsule-ext';
+
+export type WorkspaceCapsules = {
+  workspacePath: string;
+  capsules: ComponentCapsule[];
+};
